+++ CHANGELOG.md
@@ -5,6 +5,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased] - ReleaseDate
 
+### Deprecated
+
+- `meta::` queries are no deprecated and can only be used in quirks mode
+
 ### Fixed
 
 - Output annotations with the namespace "annis" in find function